Fix `node_api_create_buffer_from_arraybuffer` implementation to match Node.js behavior
