+++ src/bun.js/bindings/NapiRef.cpp
@@ -1,0 +1,46 @@
+#include "root.h"
+#include "napi.h"
+#include <wtf/TZoneMallocInlines.h>
+
+namespace Zig {
+
+WTF_MAKE_TZONE_ALLOCATED_IMPL(NapiRef);
+
+void NapiRef::ref()
+{
+    NAPI_LOG("ref %p %u -> %u", this, refCount, refCount + 1);
+    ++refCount;
+    if (refCount == 1 && !weakValueRef.isClear()) {
+        auto& vm = globalObject.get()->vm();
+        strongRef.set(vm, weakValueRef.get());
+
+        // isSet() will return always true after being set once
+        // We cannot rely on isSet() to check if the value is set we need to use isClear()
+        // .setString/.setObject/.setPrimitive will assert fail if called more than once (even after clear())
+        // We should not clear the weakValueRef here because we need to keep it if we call NapiRef::unref()
+        // so we can call the finalizer
+    }
+}
+
+void NapiRef::unref()
+{
+    NAPI_LOG("unref %p %u -> %u", this, refCount, refCount - 1);
+    bool clear = refCount == 1;
+    refCount = refCount > 0 ? refCount - 1 : 0;
+    if (clear && !m_isEternal) {
+        // we still dont clean weakValueRef so we can ref it again using NapiRef::ref() if the GC didn't collect it
+        // and use it to call the finalizer when GC'd
+        strongRef.clear();
+    }
+}
+
+void NapiRef::clear()
+{
+    NAPI_LOG("ref clear %p", this);
+    finalizer.call(env, nativeObject);
+    globalObject.clear();
+    weakValueRef.clear();
+    strongRef.clear();
+}
+
+}
